Add audio device and rate selection dialog
Lets one select the sound device and the rate.

All devices appear in the list, but non-suitable ones will be grayed
out.

There is a predefined list of common rates and non-supported rates for
the selected device will be grayed out.

The rate combo box allows typing in a rate not in the list and that rate
will be tried.  This design allows for any rate, but doesn't waste too
much time querying the audio devices for all possible rates on startup.

If setting the new device/rate doesn't work, it will automatically go
back to the previous values.

The dialog is modal.  It's necessary to select new settings and hit OK
or Cancel beforing going back to the app.

Mask the dialog handler when the rate list is populated so callbacks
aren't run as new items as added.