Add interface for querying audio device list
A new exported interface, get_audio_device(), get_audio_devices(), and
set_audio_device(), is added to audio.c.

It allows querying what devices are present, what sample rates they
support, what device is active, and changing the active device and rate.

Portaudio's ALSA driver won't let you know what rates are allowed if the
device is currently active, due to the way it tries to re-open the
device to query the rate.

This also causes a problem querying rates on a direct hardware device,
e.g. "hw:0,0", if a virtual device, e.g. "pulse" or "dmix", is active
and using that hardware.

So the common use case I want to do, where "pulse" is the default on
startup and then I want to switch to the direct hardware device to avoid
resampling, and get lower latency and better time stamping of audio,
doesn't work.  Because the hardware device is "in use" through pulse.

To get around this, a predefined list of common rates are queried on
start up and the results cached.  This way it's not necessary to query
devices after audio capture has started.