python: Allow Python code to be called from C
Add framework for being able to call Python code and write functions,
such as plot generation, in Python.

Initialize the Python interpreter.

Provide a Python extension written in C that allows Python code to
interact with tg, e.g. get audio and time data.

There is Makefile logic to turn a Python source file into C code so it
can be compiled into the binary.  It needs sed.

Framework to define a Python module's interface in C, then use those
definitions to load the code and find the methods in it.