Python interpreter
Add initial Python code.  It can:
Initialize the Python interpreter
Provide a module to python that is written in C for interacting with tg
from module writte in Python, e.g. get audio and time data.
Compile Python source files into the executable
Run the compiled in Python source modules
Get methods from the Python modules and call them.

There is Makefile logic to turn a Python source file into C code so it
can be compiled into the binary.  It needs sed.

There is C code that can create a Gdk pixmap from a Python memoryview
with copying the data and correctly handles the reference counting.
This can be used to create a GtkImage diretly from the matplotlib Agg
canvas.