Plotting of spectrograms
This uses libtfr to create time and frequency reassigned spectrograms.

One plot will be just one beat, corresponding in duration to the
waveforms tg displays, showing the drop, which is the point tg measures,
and also the average time for the unlock.  I don't see how to find which
of the tick or tock unlock time to use for a given beat, so both are
drawn.

Another plot contains the most recent N seconds of audio, which can be
controlled by a spin button.  Marks are made for where tg detected
beats.  Because the time processor is behind the audio capture, the most
recent few hundred milliseconds might not have been analyized for beats
yet.