Add audio functions for data access
These will copy requested data out of the audio buffer.  It's possible
to request by an absolute timestamp or get the most current audio.

Getting the current timestamp and then requesting audio from that time
wouldn't be the same, as new audio is added asynchronously by the
PortAudio thread.  So the timestamp would be out of date.

These can be used by the python module to get access to audio data.  Or
could be used for a playback or audio save function.