python: Code for tg extension module for Python
This creates a module named "tg" with a number of methods.  It's
implemented in C and can be called from Python.  The methods mostly
allow getting audio data and other related values.

There is something for getting the Gtk theme color to help make graphs
blend into the GUI better.