python: Code for graphing biquad filters
This generates some simple graphs for biquad filters, like the high-pass
filter tg uses on audio.

Add the Python code and the necessary calls to load it and get handles
on the methods to be used.

Limit on filter is -80 dB, not very interesting below that

Nothing will actually call the methods yet.