Keep track of half-cycle sign, i.e. tic vs toc
When finding events, keep track of which vibration we are finding:
either the tic waveform or the toc waveform.

Store events in the initial list from locate_events and later in the
processing buffers in a struct that contains a timestamp and a boolean,
which is used to indicate which vibration the event is for.

While we could use the high bit of the 64-bit timestamp or make it a signed
value and use the sign, I didn't want to always need to mask off the
timestamp or take absolute values when the data is used.