Copy the tictoc data out of processing buffer and into the snapshot
Instead of using an struct with timestampp and boolean, use two
different arrays.

This is more space efficient due to padding in the struct.  The snapshot
has a much larger list of events than an individual processing buffer so
the wasted space would matter more here.

There's also more code using the event list and it's easier to not have
to change it to use a struct.