filter: Create system for biquad filter cascade
Replace the single high pass filter with a cascade of any number of
biquadratic IIR filters, such as high pass, low pass, peaking, band
pass, notch, etc.

The system is designed to be efficient for multi-threaded operation.
The chain can be edited while the real time audio thread uses it while
minimizing when and how long the audio thread might be blocked.

To do this, the rules for accessing the chain are asymmetric.  The
thread using the chain to process audio and the thread which wants to
read and modify the filters have different rules for what they can do
with and without holding a mutex.