Horizontally flip paperstrip in vertical layout
This way the graph matches what a Vibrograf paper strip looks like.  A
movement running slow will move to the left.

The horizontal layout is unchanged.  A slower movement will move down.
This is what the Weishi 1000/1900 do, which appears to be copied from
Witschi timegraphers.

Buttons need to be adjusted to reflect this change, so their arrows
point the way their signs move.