Scrollbars for the paperstrip
Add a scrollbar for the paperstrip.  The paperstrip doesn't actually get
longer and become scrollable yet.  It only has a scrollbar in the time
dimension.

This changes the paperstrip from a GtkDrawingArea to a GtkLayout, since
the latter supports scrolling.  Then it's added into a
GtkScrolledWindow.

The scrollbar is set to not overlay onto the paperstrip.  This way it
doesn't cover it up.  Also Gtk would cause many paperstrip redraws every
time the mouse enters or leaves the window as it fades the scrollbar in
and out.  It's costly to keep redrawing the paperstrip this way and also
visually annoying as the refresh rate of the strip appears to speed up
and slow down.