Have paperstrip grow to contain all events, so it can scroll
Make the paperstrip Layout widget grow in size to cover all events in
the event list.  Since it's in a scrolled window, what this does allow
the scrollbars to scroll the Layout.

The paperstrip drawing code already gets the position of the scrollbars
to control where it gets events from.

To figure out how long the paperstrip is, the computer will find the age
of the oldest event from "now" as it finishes computing the snapshot
data.  The UI refresh will update the size of the paperstrip widget
based on this event timespan before it queues a redraw.

This isn't exactly right, as the event timespan will be a little longer
at the redraw time, since the redraw happens a little after the snapshot
to be redrawn is computed.  But we can not change the widget size at the
start of the redraw, as any change to the size causes Gtk to immediatly
requeue another redraw.  This small delay isn't really a problem and
only causes the scrollback to be maybe one pixel less than it could be.