python: Method to get last event and show event type in spectrogram
Add a Python method to get the last event time and also the sign, tic or
toc, of the event.

Buttons for tic vs toc event spectrograms

Instead of just getting whichever one was most revent, make it possible
to choose which type.

Put in the title of the plot if it's a Tic or Toc that was plotted.