python: Code for getting images from Python into Gtk
Create a Gdk Pixmap from a Python memoryview of a pixel buffer.  This is
done without copying the data and correctly handles the reference
counting.  This can be used to create a GdkPixmap directly from the
matplotlib Agg canvas.

Use this to write code for getting the canvas into a GtkImage.