filter: Audio filter chain editing dialog
Create fancy dialog that for construction of audio filter chains.  It
ended up being more complex than the rest of the app.

One can add notch filters to remove noise at specific frequencies or
peaking filters to accentuate the part of the spectrum that the watch
tick occupies.  The spectrogram function is useful to see where the
noise is.

The filter values can be adjusted with sliders and the response graph
will adjust in real time.

It's possible to see the response of just one filter or the entire
chain.

Filters can be moved around in order by drag and drop.  This was a
massive pain to implement.  I don't think Gtk uses list drag and drop
much.