filter: Get audio dialog working better with filter chain
This makes the HPF slider non-sensitive when a chain is in use that
doesn't begin with a high-pass filter.  This audio dialog only handles
that case.  More complex filters need to use the filter chain dialog.

The HFP in the audio dialog will update when it's shown, in case the
filter was edited in the filter chain dialog.

The audio dialog will emit a signal when it changes the HPF, so that the
filter dialog can update.