tppm: Run the tppm calculation in audio thread
When enabled, the audio callback will perform the TPPM calculation.
There is a new audio function to enable/diable this and another function
to get the peak value.

To be more efficient, and to deal with data alignment issues, the audio
thread will batch up audio data into a buffer until there is a certain
amount, then apply the tppm calculation to it to the find the peak from
the buffer.  The tppm module handles the buffering.

It's currently run in 100 ms chunks.  There's no advantage to using a
smaller chunk size smaller than the upgrade rate of the GUI, which is
100ms.

We need to copy the audio before the high-pass and other filters are
applied to get the true peak level as it relates to clipping in the ADC.
This is why the data in the audio ring buffer can't be used.