Speed up compute_phase
This version is about twice as fast.

Mainly this is done by swapping which is the inner most loop.  The i
loop, over each sample in a period, is ~9600 iterations and the j loop,
over each period in the buffer, is about ~80 iterations.  It's faster to
have the loop with more iterations the innermost.

This also processes the samples sequentially, instead of with a stride
of the period size, which is much more efficient.

Another speedup is that the floating point period times cycle count to
an interger compution, round(i + j *period), only needs to be done once
per cycle instead of every time, i.e. ~9600 times less often.

And then a little bit of speedup by converting the period size to an
interger when it's used as a loop limit.  I.e., i < period is
effectively (double)i < period, which is slower than i < (int)period.
The compiler doesn't make this optimization and converts i to floating
point each iteration to make a floating point compare.