Overlay amplitude graph onto paperstrip
New arrays in the snapshot will keep track of the amplitude measurements
and the time those measurements were taken.  The measurement is assigned
to the center of the averaging interval that produced it.

Since the measurement is both the tick and tock combined, they are only
recorded once per period; about half as often as the tick/tock events.

Measurements are only recorded taken when the signal level is at full
strength, i.e. the largest averaging interval is used.

It's graphed onto the paperstrip in yellow with a range of 135° to 360°,
the same as what is considered a valid amplitude.