Save averaging interval length in processing buffer
This is the same as the buffer length (sample_count) at first.  But when
the buffer gets cloned after the computer thread is done with it, only
one period of audio data is retained.  The sample_count decreases to one
period's worth, but interval_count stays the same so code can get the
averaging interval that was used.