Faster math for compute_waveform() phase
Round the float phase to an interger first, then do the phase modulus
math with intergers.  The other way around, using a floating point
modulus and then rounding later, is slower.

This also fixes a bug when the phase will be rounded up and the last
offset before it wraps is processed.  It should wrap around, but
doesn't.  E.g., suppose wf_size = 5 and phase is 2.6.  The 5 elements of
the waveform should be handled in the order:
3 4 0 1 2

But what would actually happen is:
3 4 5 1 2

Element 5 should be in the next period.  What's happening is phase+3 =
4.6, which doesn't wrap around to 0, and is then rounded up to 5, which
should wrap around, but because the modulus was done before rounding, it
doesn't.