Fix bug in count_events() with one event
If there was exactly one event in the middle of the buffer it wasn't
counted right.

This wasn't apparent at first because the 1st beat or amplitude is
always at the start and not in middle.  They can be added in the middle
when the history is cleared, but the following beats are always added in
pairs in this case, so there is never one.  But there can be one
amplitude added in the middle after a history clear.