Use union for type punning in Windows broken printf code
The code to turn a double into the raw bits typed punned through a
pointer cast, which breaks strict aliasing.  Instead use a union, which
also breaks aliasing rules, but compilers all support doing it this way.

This code is just used in MINGW because apparently printf %a doesn't
work there.